RUMM-837 Move 'batching' package into 'domain'
