+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/Reader.kt
@@ -6,7 +6,7 @@
 
 package com.datadog.android.core.internal.data
 
-import com.datadog.android.core.internal.domain.Batch
+import com.datadog.android.core.internal.data.file.Batch
 import com.datadog.tools.annotation.NoOpImplementation
 
 /**
