--- dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/batching/migrators/MoveDataMigrator.kt
@@ -1,16 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.core.internal.data.batching.migrators
-
-internal class MoveDataMigrator(
-    val pendingFolderPath: String,
-    val approvedFolderPath: String
-) : BatchedDataMigrator {
-    override fun migrateData() {
-        // TODO RUMM-838
-    }
-}
