+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/file/Batch.kt
@@ -1,0 +1,15 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.core.internal.data.file
+
+/**
+ * Represent a batch of logs read from a persisted location.
+ */
+internal data class Batch(
+    val id: String,
+    val data: ByteArray
+)
