+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/domain/batching/MigratorFactory.kt
@@ -4,9 +4,9 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.core.internal.data.batching
+package com.datadog.android.core.internal.domain.batching
 
-import com.datadog.android.core.internal.data.batching.migrators.BatchedDataMigrator
+import com.datadog.android.core.internal.domain.batching.migrators.BatchedDataMigrator
 import com.datadog.android.privacy.TrackingConsent
 
 internal interface MigratorFactory {
