+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/domain/batching/processors/DataProcessor.kt
@@ -4,12 +4,12 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.core.internal.domain
+package com.datadog.android.core.internal.domain.batching.processors
 
-/**
- * Represent a batch of logs read from a persisted location.
- */
-internal data class Batch(
-    val id: String,
-    val data: ByteArray
-)
+import com.datadog.tools.annotation.NoOpImplementation
+
+@NoOpImplementation
+internal interface DataProcessor<T : Any> {
+
+    fun consume(event: T)
+}
