+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/domain/batching/DefaultConsentAwareDataHandlerTest.kt
@@ -4,10 +4,10 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.core.internal.data.batching
+package com.datadog.android.core.internal.domain.batching
 
-import com.datadog.android.core.internal.data.batching.migrators.BatchedDataMigrator
-import com.datadog.android.core.internal.data.batching.processors.DataProcessor
+import com.datadog.android.core.internal.domain.batching.migrators.BatchedDataMigrator
+import com.datadog.android.core.internal.domain.batching.processors.DataProcessor
 import com.datadog.android.core.internal.privacy.ConsentProvider
 import com.datadog.android.privacy.TrackingConsent
 import com.datadog.android.utils.forge.Configurator
