+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/domain/batching/DefaultMigratorFactoryTest.kt
@@ -4,11 +4,11 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.core.internal.data.batching
+package com.datadog.android.core.internal.domain.batching
 
-import com.datadog.android.core.internal.data.batching.migrators.MoveDataMigrator
-import com.datadog.android.core.internal.data.batching.migrators.NoOpBatchedDataMigrator
-import com.datadog.android.core.internal.data.batching.migrators.WipeDataMigrator
+import com.datadog.android.core.internal.domain.batching.migrators.MoveDataMigrator
+import com.datadog.android.core.internal.domain.batching.migrators.NoOpBatchedDataMigrator
+import com.datadog.android.core.internal.domain.batching.migrators.WipeDataMigrator
 import com.datadog.android.privacy.TrackingConsent
 import fr.xgouchet.elmyr.Forge
 import java.util.stream.Stream
