+++ dd-sdk-android/src/test/kotlin/com/datadog/android/utils/forge/BatchForgeryFactory.kt
@@ -6,7 +6,7 @@
 
 package com.datadog.android.utils.forge
 
-import com.datadog.android.core.internal.domain.Batch
+import com.datadog.android.core.internal.data.file.Batch
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
